Eliminated END event node and instead added a check on the clock
Now the end of the simulation happens once we reach a renewal point and the clock is above End_time.
Also, minor graphical tweaks.
NOTE: The current print is just about right for 10 events (10 and not 11 since the END does not exist anymore) on the terminal of a MacBook 15" full screen. Need to be of adaptable lenght.